Remove RazorProject and move members into RazorProjectFileSystem
